Create DS1302_Demo.py
The test.py example is fine for use at the Micropython REPL prompt, but a light weight Micro:bit user may prefer a demo that can be uploaded and run directly on the Micro:bit board using the 5x5 LED matrix display's scrolling text feature to make the date and time
readable by the user. So, I wrote this.